Finally separated event types in frontend.
