+++ web-gui/src/app/rsocket/event/ClassroomEvent.ts
@@ -1,6 +1,3 @@
-import {UserCredentials} from "../../model/User";
-import {ConferenceInfo} from "../../model/ConferenceInfo";
-
 export interface ClassroomDependent {
   classroomId: string | undefined
 }
@@ -10,40 +7,5 @@ export abstract class ClassroomEvent {
 
   protected constructor(eventName: string) {
     this.eventName = eventName
-  }
-}
-
-export class MessageEvent extends ClassroomEvent {
-  message: string
-
-  constructor(message: string) {
-    super("MessageEvent")
-    this.message = message
-  }
-}
-
-export enum ConferenceAction {
-  CREATE = "CREATE",
-  CLOSE = "CLOSE",
-  VISIBILITY = "VISIBILITY",
-  USER_CHANGE = "USER_CHANGE"
-}
-
-export class ConferenceEvent extends ClassroomEvent {
-  conferenceInfo: ConferenceInfo | undefined
-  conferenceAction: ConferenceAction | undefined
-
-  constructor() {
-    super("ConferenceEvent");
-  }
-}
-
-export class InvitationEvent extends ClassroomEvent {
-  inviter: UserCredentials | undefined
-  invitee: UserCredentials | undefined
-  conferenceInfo: ConferenceInfo | undefined
-
-  constructor() {
-    super("InvitationEvent");
   }
 }
