+++ web-gui/src/app/rsocket/event/InvitationEvent.ts
@@ -1,0 +1,13 @@
+import {UserCredentials} from "../../model/User";
+import {ConferenceInfo} from "../../model/ConferenceInfo";
+import {ClassroomEvent} from "./ClassroomEvent";
+
+export class InvitationEvent extends ClassroomEvent {
+    inviter: UserCredentials | undefined
+    invitee: UserCredentials | undefined
+    conferenceInfo: ConferenceInfo | undefined
+
+    constructor() {
+        super("InvitationEvent");
+    }
+}
